requirements.txt
asttokens==2.2.1
boto3==1.26.133
botocore==1.29.133
datetime
pandas==1.5.3
plotly==5.15.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
python-dotenv==1.0.0
reportlab==4.0.4
requests==2.31.0
secure-smtplib==0.1.1
openpyxl==3.1.2 
kaleido==0.2.1
ipywidgets==8.0.6
psutil==5.9.5
streamlit==1.25.0
watchdog==3.0.0